+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/CameraStabSettings.java
@@ -173,7 +173,7 @@ public class CameraStabSettings extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			CameraStabSettings obj = new CameraStabSettings();
@@ -187,13 +187,13 @@ public class CameraStabSettings extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public CameraStabSettings GetInstance(UAVObjectManager objMngr, int instID)
+	public CameraStabSettings GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (CameraStabSettings)(objMngr.getObject(CameraStabSettings.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0x3B95DDBA;
+	protected static final long OBJID = 0x3B95DDBA;
 	protected static final String NAME = "CameraStabSettings";
 	protected static String DESCRIPTION = "Settings for the @ref CameraStab mmodule";
 	protected static final boolean ISSINGLEINST = 1 == 1;
