+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/FirmwareIAPObj.java
@@ -181,7 +181,7 @@ public class FirmwareIAPObj extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			FirmwareIAPObj obj = new FirmwareIAPObj();
@@ -195,13 +195,13 @@ public class FirmwareIAPObj extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public FirmwareIAPObj GetInstance(UAVObjectManager objMngr, int instID)
+	public FirmwareIAPObj GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (FirmwareIAPObj)(objMngr.getObject(FirmwareIAPObj.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0x3CCDFB68;
+	protected static final long OBJID = 0x3CCDFB68;
 	protected static final String NAME = "FirmwareIAPObj";
 	protected static String DESCRIPTION = "Queries board for SN, model, revision, and sends reset command";
 	protected static final boolean ISSINGLEINST = 1 == 1;
